+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/repositories/ConsumerRepository.java
@@ -18,4 +18,4 @@ package com.expediagroup.streamplatform.streamregistry.core.repositories;
 import com.expediagroup.streamplatform.streamregistry.model.Consumer;
 import com.expediagroup.streamplatform.streamregistry.model.keys.ConsumerKey;
 
-public interface ConsumerRepository extends CachedCrudRepository<Consumer, ConsumerKey> {}
+public interface ConsumerRepository extends CachingJpaRepository<Consumer, ConsumerKey> {}