+++ core/src/main/java/com/expediagroup/streamplatform/streamregistry/core/repositories/InfrastructureRepository.java
@@ -18,4 +18,4 @@ package com.expediagroup.streamplatform.streamregistry.core.repositories;
 import com.expediagroup.streamplatform.streamregistry.model.Infrastructure;
 import com.expediagroup.streamplatform.streamregistry.model.keys.InfrastructureKey;
 
-public interface InfrastructureRepository extends CachedCrudRepository<Infrastructure, InfrastructureKey> {}
+public interface InfrastructureRepository extends CachingJpaRepository<Infrastructure, InfrastructureKey> {}