+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Instrumented Resolvers with timers in line with queries and mutations
 - Enabled hibernate metrics by default
 
+## Changed
+- Utilise JPA `Example` to improve the performance of search operations in GraphQL Resolvers. 
+
 ## [0.10.3] 2020-03-17
 ### Added
 - Cache hints